+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.11.0 - 5/18/25
+## 0.11.0 - 5/19/25
 
 - Fix AS syntax highlighting
 - Fix hotreloading level assets
@@ -6,6 +6,8 @@
 - Bugfix with Twine Sugarcube not transpiling parenthesis
 - Marker api for checking a player's past choices
 - Devinput console for executing functions
+- Levels must return markers in advance via `usedMarkers()` and `grantedMarkers()`
+- Change `Periodic` class to fire immediately
 
 ## 0.10.0 - 5/5/25